server/requirements.txt
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
supabase-py==2.3.0
python-jose==3.3.0
